+++ scripts/preprocessing/_04b_run_ssp.py
@@ -60,6 +60,8 @@ def run_ssp(*, cfg, subject, session=None):
 
     ecg_projs = []
     ecg_epochs = create_ecg_epochs(raw)
+    if cfg.ssp_meg == 'auto':
+        cfg.ssp_meg = 'combined' if cfg.use_maxwell_filter else 'separate'
     if len(ecg_epochs) >= config.min_ecg_epochs:
         if cfg.ssp_reject_ecg == 'autoreject_global':
             reject_ecg_ = config.get_ssp_reject(
@@ -144,6 +146,7 @@ def get_config(
         n_proj_eog=config.n_proj_eog,
         n_proj_ecg=config.n_proj_ecg,
         ssp_meg=config.ssp_meg,
+        use_maxwell_filter=config.use_maxwell_filter,
     )
     return cfg